Two Click Navigation and Hover Comparison
You can hover over a result image and an enlarged version will be placed next to the last-clicked image for comparison.

Clicking a result or a query image will no longer navigate to that image's query. Instead, clicking "pins" the image on the enlarged board, and the user must click on another link to navigate to the image's query.